refactor(web): switch from parcel to esbuild
This means we no longer can use babel which should be fine nowadays, as
the major browsers all support ES6 fine. The build process is much
simpler now and is neatly handled using make, additionally we can get
rid of babel which brought a lot of dependencies.

Another small change: We can forgo installing the images bundled with
vis-network, as it inlines them as data URLs into the distributed CSS.

Since we provide esbuild via Nix, we don't include it in package.json.