chore(web): switch to napalm from yarn2nix
These days, support for npm's package-lock.json seems to besser than for
yarn, so let's drop yarn2nix which is not quite where it needs to be
yet (we pinned an experimental branch after all) and use napalm instead.
For this we have to migrate from yarn to npm completely.

Using npmlock2nix would be preferrable (as it doesn't require building a
Haskell registry server), but there we end up running into an unsolved
issue [1].

[1]: https://github.com/svanderburg/node2nix/issues/106